unreal_asset: still an i32 inside
unreal_asset: fix merging mistake :/

+++ unreal_modintegrator/src/handlers/ue4_23/persistent_actors.rs
@@ -490,7 +490,7 @@ pub fn handle_persistent_actors(
             actor_template.properties = determined_prop_data;
             asset.exports.push(actor_template.into());
 
-            let exports_len = PackageIndex::new(asset.exports.len());
+            let exports_len = PackageIndex::new(asset.exports.len() as i32);
             let level_export = cast!(Export, LevelExport, &mut asset.exports[level_export_index])
                 .expect("Corrupted memory");
             level_export.actors.push(exports_len);
